refactor: Attack 관련 Message를 상수로 분류
